backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
openai
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]==1.7.4 
bcrypt==4.0.1
python-multipart
